The data used in this project was obtained from the following source
 **Dataset Name**: Heart Attack Analysis & Prediction Dataset
- **Source**: https://www.kaggle.com/rashikrahmanpritom/heart-attack-analysis-prediction-dataset 

The dataset contains the age, sex, and various conditions and factors determining possibility of heart attack, which is noted using the target value as 0 (less chance of heart attack), or 1 (more chance of heart attack). It is important to note that the use of this dataset is subject to the terms and conditions provided by the data source. Please refer to the source for more information about data usage and licensing.